+++ pkgs/stopwatch/stopwatch.go
