+++ pkgs/tui/views.go
