+++ pkgs/util/util.go
@@ -1,11 +1,18 @@
 package util
 
-import ()
-
 func Map[T any, U any](slice []T, f func(T) U) []U {
 	result := make([]U, len(slice))
 	for i, v := range slice {
 		result[i] = f(v)
 	}
 	return result
+}
+
+func IndexOf(slice []string, value string) int {
+	for i, item := range slice {
+		if item == value {
+			return i
+		}
+	}
+	return -1
 }
